Do state manager fixup for for more single queries
Issue #5674

The state manager was assuming the query would take care of fixup for all single queries. However, this is only true when the relationships are explicitly included in the query. For some cases such as self references and multiple relationships between entity types, fixup is needed that will not be performed by query. Therefore we will now drop out of single query mode for any query with self refs and any time multiple entities are returned. This will have a perf impact, but super simple queries will still run in single query mode. Post RTM we will try to do something better.